+++ package.json
@@ -23,9 +23,14 @@
     "migration:generate": "npm run typeorm -- migration:generate"
   },
   "dependencies": {
+    "@bull-board/api": "^6.9.2",
+    "@bull-board/express": "^6.9.2",
+    "@bull-board/nestjs": "^6.9.2",
+    "@nestjs/bull": "^11.0.2",
     "@nestjs/common": "^9.3.9",
     "@nestjs/config": "^2.3.1",
     "@nestjs/core": "^9.3.9",
+    "@nestjs/event-emitter": "^3.0.1",
     "@nestjs/jwt": "^10.0.2",
     "@nestjs/microservices": "^9.3.9",
     "@nestjs/mongoose": "^9.2.1",
@@ -46,10 +51,12 @@
     "axios": "^1.3.6",
     "base-64": "^1.0.0",
     "bcrypt": "^5.1.1",
+    "bull": "^4.16.5",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
     "cross-env": "^7.0.3",
     "dayjs": "^1.11.10",
+    "express-basic-auth": "^1.2.1",
     "fast-xml-parser": "^4.2.5",
     "form-data": "^4.0.0",
     "getmac": "^6.6.0",
