+++ src/invoice/dto/query-result-invoice.ts
@@ -3,6 +3,7 @@ import { InvoiceEntity } from '../entities/invoice.entity';
 export class QueryResultInvoice {
   invoice: InvoiceEntity;
   fileName: string;
+  message: string;
   codigo_respuesta_sunat: string;
   codigo_respuesta_sunat_int: number;
   documento: string;
@@ -13,4 +14,5 @@ export class QueryResultInvoice {
   xml: string;
   cdr: string;
   pdfA4: string;
+  sendMode: string;
 }
