+++ src/invoice/invoice.module.ts
@@ -1,4 +1,5 @@
-import { Module } from '@nestjs/common';
+import { BullBoardModule } from '@bull-board/nestjs';
+import { Module, forwardRef } from '@nestjs/common';
 import { InvoiceService } from './services/invoice.service';
 import { InvoiceController } from './controllers/invoice.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
@@ -15,15 +16,25 @@ import { EntidadEntity } from 'src/entidades/entities/entidad.entity';
 import { UserEntity } from 'src/user/entities/user.entity';
 import { SeriesEntity } from 'src/series/entities/series.entity';
 import { ConfiguracionesModule } from 'src/configuraciones/configuraciones.module';
-import { InvoiceGateway } from './gateway/invoice.gateway';
 import { ConnectionModule } from 'src/connection/connection.module';
-import { JwtStrategy } from 'src/lib/strategies/jwt.strategies';
-import { AuthModule } from 'src/auth/auth.module';
-import { ResourcesUsersModule } from 'src/resources-users/resources-users.module';
 import { InvoiceDetailsEntity } from './entities/invoice_details.entity';
 import { AnulacionEntity } from 'src/anulaciones/entities/anulacion.entity';
 import { CodesReturnSunatModule } from 'src/codes-return-sunat/codes-return-sunat.module';
 import { PosModule } from 'src/pos/pos.module';
+import { SunatService } from 'src/sunat/services/sunat.service';
+import { GatewayModule } from 'src/gateway/gateway.module';
+import { InvoiceProcessor } from 'src/invoice/processors/invoice.processor';
+import { BullModule } from '@nestjs/bull';
+import { BullAdapter } from '@bull-board/api/bullAdapter';
+import { SocketSessionModule } from 'src/socket-session/socket-session.module';
+import { InvoiceManualJobs } from './jobs/invoice-manual.job';
+import { InvoiceScheduledJobs } from './jobs/invoice-scheduled.job';
+import { SunatScheduledProcessor } from './processors/sunat-scheduled.processor';
+import { SunatManualProcessor } from './processors/sunat-manual.processor';
+import { SunatInmediateProcessor } from './processors/sunat-inmediate.processor';
+import { SunatPendingProcessor } from './processors/sunat-pending.processor';
+import { InvoiceInmediateJobs } from './jobs/invoice-inmediate.job';
+import { InvoicePendingJobs } from './jobs/invoice-pending.job';
 
 @Module({
   imports: [
@@ -45,13 +56,64 @@ import { PosModule } from 'src/pos/pos.module';
     TipodocsModule,
     ConfiguracionesModule,
     ConnectionModule,
-    AuthModule,
-    ResourcesUsersModule,
     CodesReturnSunatModule,
     PosModule,
+    forwardRef(() => GatewayModule),
+    BullModule.registerQueue(
+      {
+        name: 'invoice-submission', // create sale and bull inmediate - scheduled
+      },
+      {
+        name: 'job-sunat-submission-inmediate', // job inmediate
+      },
+      {
+        name: 'job-sunat-submission-pending', // job pending
+      },
+      {
+        name: 'job-sunat-submission-manual', // job manual
+      },
+      {
+        name: 'job-sunat-submission-scheduled', // job scheduled
+      },
+    ),
+    BullBoardModule.forFeature(
+      {
+        name: 'invoice-submission',
+        adapter: BullAdapter, //or use BullAdapter if you're using bull instead of bullMQ
+      },
+      {
+        name: 'job-sunat-submission-inmediate',
+        adapter: BullAdapter, //or use BullAdapter if you're using bull instead of bullMQ
+      },
+      {
+        name: 'job-sunat-submission-pending',
+        adapter: BullAdapter, //or use BullAdapter if you're using bull instead of bullMQ
+      },
+      {
+        name: 'job-sunat-submission-manual',
+        adapter: BullAdapter, //or use BullAdapter if you're using bull instead of bullMQ
+      },
+      {
+        name: 'job-sunat-submission-scheduled',
+        adapter: BullAdapter, //or use BullAdapter if you're using bull instead of bullMQ
+      },
+    ),
+    SocketSessionModule,
   ],
   controllers: [InvoiceController],
-  providers: [InvoiceService, InvoiceGateway, JwtStrategy],
+  providers: [
+    InvoiceService,
+    SunatService,
+    InvoiceProcessor,
+    SunatInmediateProcessor,
+    SunatManualProcessor,
+    SunatScheduledProcessor,
+    SunatPendingProcessor,
+    InvoiceInmediateJobs,
+    InvoiceManualJobs,
+    InvoiceScheduledJobs,
+    InvoicePendingJobs,
+  ],
   exports: [InvoiceService],
 })
 export class InvoiceModule {}
