+++ src/lib/enum/envio_sunat_modo.enum.ts
@@ -1,0 +1,6 @@
+export enum EnvioSunatModo {
+  INMEDIATO = 'INMEDIATO', // Envío inmediato
+  MANUAL = 'MANUAL', // Envio manual
+  PROGRAMADO = 'PROGRAMADO', // Envío en un tiempo específico
+  NO_ENVIA = 'NO_ENVIA', // No envia
+}
