+++ src/lib/strategies/jwt.strategies.ts
@@ -96,9 +96,10 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
                 estado: item.estado,
                 configuraciones: item.configsEmpresa,
                 establecimientos: item.establecimientos.map((est) => {
-                  const { empresa, ...rest } = est;
+                  const { configsEstablecimiento, empresa, ...rest } = est;
                   return {
                     ...rest,
+                    configuraciones: est.configsEstablecimiento,
                     codigo: est.codigo === '0000' ? 'PRINCIPAL' : est.codigo,
                   };
                 }),
@@ -122,9 +123,10 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
                   establecimiento: empresa.establecimientos
                     .filter((establecimiento) => establecimiento.estado)
                     .map((est) => {
-                      const { empresa, ...rest } = est;
+                      const { configsEstablecimiento, empresa, ...rest } = est;
                       return {
                         ...rest,
+                        configuraciones: est.configsEstablecimiento,
                         codigo:
                           est.codigo === '0000' ? 'PRINCIPAL' : est.codigo,
                         pos:
