+++ src/sunat/dto/generate-xml-invoice.dto.ts
@@ -1,0 +1,5 @@
+export class GenerateXmlInvoiceDto {
+  xmlUnsigned: string;
+  fileName: string;
+  fileNameExtension: string;
+}
